Move api error and codes from chain module to http_api
