+++ framework/src/modules/chain/modules/signatures.js
@@ -14,8 +14,6 @@
 
 'use strict';
 
-const apiCodes = require('../helpers/api_codes.js');
-const ApiError = require('../helpers/api_error.js');
 const Signature = require('../logic/signature.js');
 const transactionTypes = require('../helpers/transaction_types.js');
 
@@ -34,8 +32,6 @@ __private.assetTypes = {};
  * @class
  * @memberof modules
  * @see Parent: {@link modules}
- * @requires helpers/api_codes
- * @requires helpers/api_error
  * @requires helpers/transaction_types
  * @requires logic/signature
  * @param {function} cb - Callback function
@@ -96,29 +92,6 @@ Signatures.prototype.onBind = function(scope) {
 	__private.assetTypes[transactionTypes.SIGNATURE].bind(scope.modules.accounts);
 };
 
-__private.processPostResult = function(err, res, cb) {
-	let error = null;
-	let response = null;
-
-	// TODO: Need to improve error handling so that we don't
-	// need to parse the error message to determine the error type.
-	const processingError = /^Error processing signature/;
-	const badRequestBodyError = /^Invalid signature body/;
-
-	if (err) {
-		error = new ApiError(err, apiCodes.PROCESSING_ERROR);
-	} else if (res.success) {
-		response = { status: 'Signature Accepted' };
-	} else if (processingError.test(res.message)) {
-		error = new ApiError(res.message, apiCodes.PROCESSING_ERROR);
-	} else if (badRequestBodyError.test(res.message)) {
-		error = new ApiError(res.message, apiCodes.BAD_REQUEST);
-	} else {
-		error = new ApiError(res.message, apiCodes.INTERNAL_SERVER_ERROR);
-	}
-	return setImmediate(cb, error, response);
-};
-
 // Shared API
 /**
  * Public methods, accessible via API.
@@ -135,9 +108,9 @@ Signatures.prototype.shared = {
 	 * @returns {setImmediateCallback} cb
 	 */
 	postSignature(signature, cb) {
-		return modules.transport.shared.postSignature({ signature }, (err, res) => {
-			__private.processPostResult(err, res, cb);
-		});
+		return modules.transport.shared.postSignature({ signature }, (err, res) =>
+			setImmediate(cb, err, res)
+		);
 	},
 
 	/**
@@ -148,11 +121,8 @@ Signatures.prototype.shared = {
 	 * @returns {setImmediateCallback} cb
 	 */
 	postSignatures(signatures, cb) {
-		return modules.transport.shared.postSignatures(
-			{ signatures },
-			(err, res) => {
-				__private.processPostResult(err, res, cb);
-			}
+		return modules.transport.shared.postSignatures({ signatures }, (err, res) =>
+			setImmediate(cb, err, res)
 		);
 	},
 };
