+++ framework/test/mocha/functional/http/post/0.transfer.js
@@ -23,7 +23,7 @@ const phases = require('../../../common/phases');
 const sendTransactionPromise = require('../../../common/helpers/api')
 	.sendTransactionPromise;
 const randomUtil = require('../../../common/utils/random');
-const errorCodes = require('../../../../../src/modules/chain/helpers/api_codes');
+const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
 const Bignum = require('../../../../../src/modules/chain/helpers/bignum.js');
 
 const { NORMALIZER } = global.constants;
