+++ framework/test/mocha/functional/http/post/4.multisignature.advanced.js
@@ -22,7 +22,7 @@ const Scenarios = require('../../../common/scenarios');
 const waitFor = require('../../../common/utils/wait_for');
 const randomUtil = require('../../../common/utils/random');
 const apiHelpers = require('../../../common/helpers/api');
-const errorCodes = require('../../../../../src/modules/chain/helpers/api_codes');
+const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
 
 const { NORMALIZER } = global.constants;
 const sendTransactionPromise = apiHelpers.sendTransactionPromise;
